Moved the old test into a test directory for Mocha, added a new test for the index.js file. Since the command line functionality is now in the index script, moved that testing from the greeter file to the index file. Tests should now match scripts in terms of content.
