+++ greeter.js
@@ -1,12 +1,5 @@
 /** Created by Gloria on 10/17/16. **/
 
-
-if (process.argv.length > 2) {
-  // if this is called with more than just node and the script name, pass those to sayHello
-  argString = process.argv.slice(2).join(' ');
-  sayHello(argString);
-}
-
 function sayHello(name) {
   var salutation = 'Hello';
